add generic token processor
